Add files via upload
Principal Component Analysis is an unsupervised learning algorithm that is used for the dimensionality reduction in machine learning.
PCA helps you interpret your data, but it will not always find the important patterns. Principal component analysis (PCA) simplifies the complexity in high-dimensional data while retaining trends and patterns. It does this by transforming the data into fewer dimensions, which act as summaries of features